requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
asqlite==2.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
braceexpand==0.1.7
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
discord.py==2.4.0
docutils==0.20.1
frozenlist==1.4.1
idna==3.7
imageio-ffmpeg==0.5.1
imagesize==1.4.1
import_expression==1.1.5
Jinja2==3.1.4
jishaku==2.5.2
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.1
pycparser==2.22
Pygments==2.18.0
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
snowballstemmer==2.2.0
snowflake-id==1.0.2
Sphinx==7.4.7
sphinx-rtd-dark-mode==1.3.0
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
starlight-dpy @ git+https://github.com/InterStella0/starlight-dpy@1829405efa50fc5d73f3d59ed02c1274dd2a5d33
tomli==2.0.1
urllib3==2.2.2
yarl==1.9.4
